logo component moved to logo folder
