+++ b2b-sample/guardio-insurance-sample-application/components/logo/logoComponent.js
@@ -17,9 +17,9 @@
  */
 
 import React from "react";
+import Logo from "./logo";
 import config from "../../config.json";
 import styles from "../../styles/Settings.module.css";
-import Logo from "../logo/logo";
 
 /**
  *
